+++ addons/yat/src/attributes/CommandInputAttribute.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using Godot;
 using YAT.Classes;
 using YAT.Helpers;
 
@@ -9,24 +8,26 @@ namespace YAT.Attributes;
 [AttributeUsage(AttributeTargets.Property, Inherited = true, AllowMultiple = false)]
 public class CommandInputAttribute : Attribute
 {
-    public StringName Name { get; private set; }
-    public List<CommandType> Types { get; private set; } = new();
-    public string Description { get; private set; }
+    public readonly string Name, Description;
+    public readonly List<CommandType> Types = new();
 
-    public CommandInputAttribute(StringName name, string type, string description = "")
+    public CommandInputAttribute(string name, string type, string description = "")
     {
         Name = name;
         Description = description;
 
         if (string.IsNullOrEmpty(type))
-            GD.PushError(Messages.InvalidCommandInputType(type, name));
+            throw new ArgumentNullException(Messages.InvalidCommandInputType(type, name));
 
         foreach (var t in type.Split('|'))
         {
             if (Parser.TryParseCommandInputType(t, out var commandInputType))
+            {
                 Types.Add(commandInputType);
-            else
-                GD.PushError(Messages.InvalidCommandInputType(t, name));
+                continue;
+            }
+
+            throw new ArgumentException(Messages.InvalidCommandInputType(t, name));
         }
     }
 }